Meta-analysis of external control studies using JAGS

Perform a meta-analysis of studies in which external control arms were compared to internal control arms using rjags. You can view the ecmeta() documentation for more general details.

ecmeta_jags(
  data,
  n_chains = 1,
  n_iter = 1000,
  thin = 1,
  n_burnin = 0,
  n_adapt = 1000,
  quiet = FALSE,
  prior_mu = normal(0, 100),
  prior_scale = invgamma(0.001, 0.001)
)

Arguments

data
A loghr_data object that stores estimates of the log hazard ratios of the internal control arms relative to the external control arms.

n_chains
The number of parallel chains for the model. Passed to rjags::jags.model().

n_iter
Number of iterations to monitor. Passed to rjags::coda.samples().

thin
Thinning interval for monitors. Passed to rjags::coda.samples().

n_burnin
Number of

n_adapt
The number of iterations for adaptation. Passed to rjags::jags.model().

quiet
If TRUE then messages generated during compilation will be suppressed, as well as the progress bar during adaptation. Passed to rjags::jags.model().

prior_mu
Prior for mu. Must be specified as a normal distribution with normal().

prior_scale
Prior for the scale parameter, which is either sigma or sigma^2 depending on the choice of prior. A half student_t() distribution or uniform prior can be placed on sigma while an invgamma() prior can be used for sigma^2.

Value

An ecmeta_jags object, which is a list containing the following elements:

mcmc

A coda::mcmc.list object containing samples of the model parameters.

data

The data object.