Predict log hazard ratio

Predict the true log hazard ratio in a new single-arm study by adjusting for additional bias and variability caused by the non-randomized design. The function first uses MCMC to sample the true log hazard ratio comparing the treatment to the external control (trt_ec). It then draws the true log hazard ratio comparing the internal control to the external control (ic_ec) using the posterior samples of $mu$ and $sigma$ stored in object. Lastly, trt_ec and ic_ec are combined to produce posterior predictive draws of the true log hazard ratio in a comparison of the treatment to the internal control (trt_ic).

# S3 method for ecmeta_jags
predict(
  object,
  new_estimate,
  new_standard_error,
  n_burnin = 0,
  thin = 1,
  n_adapt = 1000,
  quiet = FALSE,
  ...
)

Arguments

object
An ecmeta_jags object containing posterior draws of .

new_estimate
An estimate of the log hazard ratio for a comparison of the treatment arm to the external control.

new_standard_error
Standard error of the log hazard ratio for a comparison of the treatment arm to the external control.

n_burnin
Number of

thin
Thinning interval for monitors. Passed to rjags::coda.samples().

n_adapt
The number of iterations for adaptation. Passed to rjags::jags.model().

quiet
If TRUE then messages generated during compilation will be suppressed, as well as the progress bar during adaptation. Passed to rjags::jags.model().

...
Currently unnused.

Value

A list containing the following elements:

mcmc

A coda::mcmc.list object containing samples of the true log hazard ratio comparing the treatment to the external control.

loghr

A matrix with three columns containing draws of the true log hazard ratios. The columns are: trt_ic (a comparison of the treatment to the internal control), trt_ec (a comparison of the treatment to the external control) and ic_ec (a comparison of the internal control to the external control).

See also