Summary of external control meta-analysis

Summary statistics for a meta-analytic model fit using ecmeta().

# S3 method for ecmeta_jags
summary(object, quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975), ...)

Arguments

object
An object of the appropriate class estimated using ecmeta().

quantiles
A vector of quantiles to evaluate for each variable.

...
Additional arguments to pass to coda::summary.mcmc.list().

Value

A data.frame with one row for each parameter (mu and sigma). The columns contain the same summary statistics and quantiles as in coda::summary.mcmc.list(). Statistics reported are the mean (mean), standard deviation (sd), naive (ignoring autocorrelation of the chain) standard error of the mean (se_mean), and time-series standard error based on an estimate of the spectral density at 0 (se_ts). Additional columns contain quantiles of the distribution as specified by quantiles. Finally, if multiple chains were used, the column rhat is included, which is the R-hat convergence diagnostic (values substantially above 1 indicate a lack of convergence).