Summary of predicted log hazard ratios

Summary statistics for true log hazard ratios (HRs) predicted using ecmeta prediction functions.

# S3 method for ecmeta_jags_prediction
summary(
  object,
  exponentiate = FALSE,
  quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  ...
)

Arguments

object
A object of the appropriate class.

exponentiate
Logical indicating whether or not to exponentiate the the log HR. Defaults to FALSE; if TRUE, then HRs are summarized.

...
Additional arguments to pass to coda::summary.mcmc.list().

Value

A data.frame with one row for each log HR comparison: treatment vs. internal control (loghr_trt_ic), treatment vs. external control (loghr_trt_ec), and internal control vs. external control (loghr_ic_ec).

The columns contain summary statistics and quantiles. Statistics reported are the mean (mean), standard deviation (sd), and naive (ignoring autocorrelation of the chain) standard error of the mean. Additional columns contain quantiles of the distribution as specified by quantiles. Finally, if multiple chains were used, the column rhat is included, which is the R-hat convergence diagnostic (values substantially above 1 indicate a lack of convergence). Note that rhat is only provided for loghr_trt_ec. Diagnostics for loghr_ic_ec should be based on mu and sigma from ecmeta() and loghr_trt_ic is deterministically related to loghr_trt_ec and loghr_ic_ec.