Summary of external control meta-analysis using maximum likelihood

Summary statistics for a meta-analytic model fit using ecmeta() with method == "ml".

# S3 method for ecmeta_ml
summary(object, level = 0.95, ...)

Arguments

object
An ecmeta_ml object.

level
The confidence level. Default is 0.95 for a 95 percent confidence interval.

...
Currently unused.

Value

A data.frame with one row for each parameter (mu and sigma). There is always one column named estimate containing the point estimates. In addition, if a variance-covariance matrix was estimated, then there are additional columns for the standard errors as well as the lower and upper confidence limits.

Note

object contains estimates of logsigma rather than sigma. Point estimates and confidence intervals for sigma are computed by simply exponentiating logsigma; the standard error for sigma is estimated using the delta method.